fixed Dashboardviewmodel UnitTests and refactored
