+++ ClearScore.xcodeproj/xcuserdata/thomasrichardson.xcuserdatad/xcdebugger/Breakpoints_v2.xcbkptlist
@@ -3,70 +3,4 @@
    uuid = "96C81A8E-0FE2-47AB-99C8-314D5C925C35"
    type = "1"
    version = "2.0">
-   <Breakpoints>
-      <BreakpointProxy
-         BreakpointExtensionID = "Xcode.Breakpoint.FileBreakpoint">
-         <BreakpointContent
-            uuid = "53F92BD5-60CC-4465-AF64-E9AD94ACD0FC"
-            shouldBeEnabled = "No"
-            ignoreCount = "0"
-            continueAfterRunningActions = "No"
-            filePath = "ClearScoreTests/Mocks/MockAPI.swift"
-            startingColumnNumber = "9223372036854775807"
-            endingColumnNumber = "9223372036854775807"
-            startingLineNumber = "14"
-            endingLineNumber = "14"
-            landmarkName = "execute(apiRequest:)"
-            landmarkType = "7">
-            <Locations>
-               <Location
-                  uuid = "53F92BD5-60CC-4465-AF64-E9AD94ACD0FC - 948949a7f7a4480f"
-                  shouldBeEnabled = "Yes"
-                  ignoreCount = "0"
-                  continueAfterRunningActions = "No"
-                  symbolName = "ClearScoreTests.MockAPI.execute&lt;&#x3c4;_0_0 where &#x3c4;_0_0: ClearScore.APIRequest&gt;(apiRequest: &#x3c4;_0_0) async throws -&gt; &#x3c4;_0_0.ResponseBody"
-                  moduleName = "ClearScoreTests"
-                  usesParentBreakpointCondition = "Yes"
-                  urlString = "file:///Users/thomasrichardson/ClearDay/ClearScore/ClearScoreTests/Mocks/MockAPI.swift"
-                  startingColumnNumber = "9223372036854775807"
-                  endingColumnNumber = "9223372036854775807"
-                  startingLineNumber = "15"
-                  endingLineNumber = "15"
-                  offsetFromSymbolStart = "393">
-               </Location>
-               <Location
-                  uuid = "53F92BD5-60CC-4465-AF64-E9AD94ACD0FC - 948949a7f7a4480f"
-                  shouldBeEnabled = "Yes"
-                  ignoreCount = "0"
-                  continueAfterRunningActions = "No"
-                  symbolName = "ClearScoreTests.MockAPI.execute&lt;&#x3c4;_0_0 where &#x3c4;_0_0: ClearScore.APIRequest&gt;(apiRequest: &#x3c4;_0_0) async throws -&gt; &#x3c4;_0_0.ResponseBody"
-                  moduleName = "ClearScoreTests"
-                  usesParentBreakpointCondition = "Yes"
-                  urlString = "file:///Users/thomasrichardson/ClearDay/ClearScore/ClearScoreTests/Mocks/MockAPI.swift"
-                  startingColumnNumber = "9223372036854775807"
-                  endingColumnNumber = "9223372036854775807"
-                  startingLineNumber = "15"
-                  endingLineNumber = "15"
-                  offsetFromSymbolStart = "394">
-               </Location>
-            </Locations>
-         </BreakpointContent>
-      </BreakpointProxy>
-      <BreakpointProxy
-         BreakpointExtensionID = "Xcode.Breakpoint.FileBreakpoint">
-         <BreakpointContent
-            uuid = "35703496-F099-4E00-B30C-03DF6E4A382C"
-            shouldBeEnabled = "No"
-            ignoreCount = "0"
-            continueAfterRunningActions = "No"
-            filePath = "ClearScore/Dashboard/DashboardViewModel.swift"
-            startingColumnNumber = "9223372036854775807"
-            endingColumnNumber = "9223372036854775807"
-            startingLineNumber = "19"
-            endingLineNumber = "19"
-            landmarkName = "fetchDashboardData()"
-            landmarkType = "7">
-         </BreakpointContent>
-      </BreakpointProxy>
-   </Breakpoints>
 </Bucket>
