+++ ClearScore/Lifecycle/AppDelegate.swift
