+++ ClearScore/Lifecycle/SceneDelegate.swift
