+++ ClearScore/Networking/API.swift
@@ -5,24 +5,3 @@ protocol APIType {
     func executeRaw<T: APIRequest>(apiRequest: T) async throws -> (data: Data, response: URLResponse)
 
 }
-
-protocol URLSessionType {
-    func data(request: URLRequest) async throws -> (Data, URLResponse)
-}
-extension URLSession: URLSessionType {
-    @available(iOS, deprecated: 15.0, message: "This extension is no longer necessary. Use API built into SDK")
-    func data(request: URLRequest) async throws -> (Data, URLResponse) {
-        try await withCheckedThrowingContinuation { continuation in
-            let task = self.dataTask(with: request) { data, response, error in
-                guard let data = data, let response = response else {
-                    let error = error ?? URLError(.badServerResponse)
-                    return continuation.resume(throwing: error)
-                }
-
-                continuation.resume(returning: (data, response))
-            }
-
-            task.resume()
-        }
-    }
-}
