+++ ClearScore/Networking/Enums/APIError.swift
@@ -1,4 +1,5 @@
 import Foundation
-public enum APIError: Error, LocalizedError {
+
+enum APIError: Error, LocalizedError {
     case invalidUrl(urlString: String)
 }
