+++ ClearScore/Networking/Enums/Headers.swift
@@ -1,5 +1,5 @@
 import Foundation
-// This should be living in a shared module
+
 enum Headers {
     static let contentType = "Content-Type"
     static let contentTypeJson = "application/json"
