+++ ClearScore/Networking/Protocols/URLSessionType.swift
@@ -1,0 +1,4 @@
+import Foundation
+protocol URLSessionType {
+    func data(request: URLRequest) async throws -> (Data, URLResponse)
+}
